Uso de configuraciones, colores de jugadores y dibujo barras de vida
Se utilizan las configuraciones iniciales obtenidas del YAML.
Se dibujan las barras de vida de cada jugador, siendo estas las sumas de las
vidas de los gusanos.

+++ src/client/configuracionCliente.h
@@ -9,7 +9,7 @@
 #include <vector>
 #include "defs.h"
 
-typedef std::vector<int> colorJugador;
+typedef std::vector<uint8_t> colorJugador;
 
 class ConfiguracionCliente {
 private:
